// packages/scheme/generated/models/RegisterCredentials.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RegisterCredentials = {
    /**
     * ユーザー名
     */
    name: string;
    /**
     * メールアドレス
     */
    email: string;
    /**
     * パスワード
     */
    password: string;
    /**
     * パスワード（確認用）
     */
    password_confirmation: string;
};